Last digit

+++ 0x01-variables_if_else_while/1-last_digit.c
@@ -12,7 +12,7 @@
 int main(void)
 {
 int n;
-int lstdgt;
+long lstdgt;
 srand(time(0));
 n = rand() - RAND_MAX / 2;
  lstdgt = n % 10;
